In order to run any JVM app on compact profile 1, 2 or 3 JRE the java bootstrap code under runjava cannot depend on java.beans.* that is not part of any of the compact profiles. More specifically runjava depends on cglib and asm java libraries that manipulate bytecode and rely on java.beans.
Therefore the changes that are part of this commit essentially allow to bootstrap a JVM app in one of two modes:
  - old one which provides classloader and JUL log manager isolation between potential multiple apps and runjava code and apps that possibly targets multi-apps on single JVM in OSv and requires full JRE
  - new one which does NOT provide any classloader and JUL log manager isolation and targets single main method apps on JVM in OSv and allows using compact profile 1, 2 or 3

Following changes are part of this commit:

* Added io.osv.isolated and io.osv.nonisolated packages under java/runjava and refactored ContextIsolator to create JvmApp, NonIsolatedJvmApp, RunNonIsolatedJvmApp, solatedJvmApp, NonIsolatedJvmApp

* Changed java.cc and makefile to support producing two executable - old java.so that can run multiple JVM apps in isolated fashion (class loader, log manager, etc) and new java_non_isolated.so that can run single JVM app without any isolation of the former one

* Changed java/jvm/java.cc to print which java class it uses to bootstrap - isolated or nonisolated one; fixed Makefile to properly handle building of java_non_isolated.so

* Added java_non_isolated.so to modules/java-tests/usr.manifest

* Added java_non_isolated test to test.py; modified testing.py to detect failure to start *.so

* Added unit tests to test running non-isolated JVM app to tests-isolates
