+++ .gitignore
@@ -15,6 +15,7 @@ cscope.po.out
 *.pyc
 *.swp
 *.patch
+*.iml
 
 *.orig
 *.rej
