+++ java/runjava/src/main/java/io/osv/nonisolated/AppThreadTerminatedWithUncaughtException.java
@@ -1,4 +1,4 @@
-package io.osv;
+package io.osv.nonisolated;
 
 /*
  * Copyright (C) 2014 Cloudius Systems, Ltd.
@@ -6,8 +6,8 @@ package io.osv;
  * This work is open source software, licensed under the terms of the
  * BSD license as described in the LICENSE file in the top-level directory.
  */
-public class ContextFailedException extends Exception {
-    public ContextFailedException(Throwable cause) {
+public class AppThreadTerminatedWithUncaughtException extends Exception {
+    public AppThreadTerminatedWithUncaughtException(Throwable cause) {
         super(cause);
     }
 }
