+++ java/runjava/src/main/java/io/osv/nonisolated/RunNonIsolatedJvmApp.java
@@ -1,0 +1,50 @@
+package io.osv.nonisolated;
+
+/*
+ * Copyright (C) 2013-2014 Cloudius Systems, Ltd.
+ *
+ * This work is open source software, licensed under the terms of the
+ * BSD license as described in the LICENSE file in the top-level directory.
+ */
+
+import io.osv.MainClassNotFoundException;
+
+public class RunNonIsolatedJvmApp {
+	private static native void onVMStop();
+
+	static {
+		Runtime.getRuntime().addShutdownHook(new Thread() {
+			public void run() {
+				onVMStop();
+			}
+		});
+	}
+
+    public static void main(String[] args) {
+        if (args.length > 0 && args[0].equals("-version")) {
+            System.err.println("java version \"" +
+                    System.getProperty("java.version") + "\"");
+            System.err.println(System.getProperty("java.runtime.name") +
+                    " (" + System.getProperty("java.runtime.version") +
+                    ")");
+            System.err.println(System.getProperty("java.vm.name") +
+                    " (build " + System.getProperty("java.vm.version") +
+                    ", " + System.getProperty("java.vm.info") + ")");
+            return;
+        }
+
+        try {
+            NonIsolatedJvmApp.getInstance().runSync(args);
+        } catch (IllegalArgumentException ex) {
+            System.err.println("RunJava: " + ex.getMessage());
+        } catch (AppThreadTerminatedWithUncaughtException ex) {
+            if (ex.getCause() instanceof MainClassNotFoundException) {
+                System.err.println("Error: Could not find or load main class " + ((MainClassNotFoundException) ex.getCause()).getClassName());
+            } else {
+                ex.printStackTrace();
+            }
+        } catch (Throwable ex) {
+            ex.printStackTrace();
+        }
+    }
+}
