+++ java/tests/src/main/java/io/osv/PropertyIsolationTest.java
@@ -6,7 +6,9 @@ import tests.PropertySetter;
 
 import java.util.concurrent.CyclicBarrier;
 
-import static io.osv.TestIsolateLaunching.runIsolate;
+import io.osv.isolated.Context;
+
+import static io.osv.TestLaunching.runIsolate;
 import static java.util.Arrays.asList;
 import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
